+++ cellmaps_coembedding/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = 'Cell Maps team'
 __email__ = 'tools@cm4ai.org'
-__version__ = '0.1.0'
+__version__ = '0.4.0'
 __repo_url__ = 'https://github.com/idekerlab/cellmaps_coembedding'
 __description__ = 'A tool to generate coembeddings from IF image embeddings and PPI network embeddings'
 __computation_name__ = 'Merged Embedding'