+++ cellmaps_coembedding/autoembed_sc/architecture.py
@@ -4,10 +4,12 @@ import numpy as np
 import torch.nn as nn
 from torch.utils.data import Dataset
 
+
 class ToTensor:
     """
     A class that converts a numpy ndarray to a torch Tensors.
     """
+
     def __call__(self, sample):
         """
         Convert the input numpy ndarray to a float Tensor.
@@ -24,6 +26,7 @@ class Modality:
     """
     Represents a single modality of data, containing training features and labels.
     """
+
     def __init__(self, training_data, name, transform, device):
         """
         Initialize the Modality object with given training data, a name, a transformation, and the device.
@@ -56,6 +59,7 @@ class Protein_Dataset(Dataset):
     """
     A dataset class for handling protein data across multiple modalities.
     """
+
     def __init__(self, wrapper):
         """
         Initialize the dataset using a dictionary of modalities.
@@ -66,7 +70,7 @@ class Protein_Dataset(Dataset):
         self.protein_dict = dict()
         self.mask_dict = dict()
         self.device = wrapper.device
-        
+
         for modality in wrapper.modalities_dict.values():
             for i in np.arange(len(modality.train_labels)):
                 protein_name = modality.train_labels[i]
@@ -85,7 +89,7 @@ class Protein_Dataset(Dataset):
                 if modality.name not in self.protein_dict[protein_name]:
                     self.protein_dict[protein_name][modality.name] = torch.zeros(modality.input_dim).to(wrapper.device)
                     self.mask_dict[protein_name][modality.name] = 0
-                    
+
         self.protein_ids = dict(zip(np.arange(len(self.protein_dict.keys())), self.protein_dict.keys()))
 
     def __len__(self):
@@ -114,6 +118,7 @@ class TrainingDataWrapper:
     """
     Wraps training data for all modalities.
     """
+
     def __init__(self, modality_data, modality_names, device, l2_norm, dropout, latent_dim, hidden_size_1,
                  hidden_size_2,
                  resultsdir, ):
@@ -170,6 +175,7 @@ class uniembed_nn(nn.Module):
     """
     A neural network model for embedding proteins using multiple modalities.
     """
+
     def __init__(self, data_wrapper):
         """
         Initialize the model using a data wrapper that contains modality data configurations.
@@ -204,10 +210,8 @@ class uniembed_nn(nn.Module):
                 nn.Linear(data_wrapper.hidden_size_1, modality.input_dim))
             #     decoder.apply(init_weights)
 
-            
             self.encoders[modality.name] = encoder
             self.decoders[modality.name] = decoder
-            
 
     def forward(self, inputs):
         """