+++ cellmaps_coembedding/cellmaps_coembeddingcmd.py
@@ -1,6 +1,7 @@
 #! /usr/bin/env python
 
 import argparse
+import json
 import os
 import sys
 import logging
@@ -47,35 +48,40 @@ def _parse_arguments(desc, args):
     parser.add_argument(IMAGE_EMBEDDINGDIR,
                         help='Directory aka rocrate image embedding '
                              'file resides (Deprecated: use --embeddings flag)')
-    parser.add_argument('--latent_dimension', type=int, default=128,
+    parser.add_argument('--latent_dimension', type=int, default=EmbeddingGenerator.LATENT_DIMENSIONS,
                         help='Output dimension of embedding')
-    parser.add_argument('--n_epochs_init', default=200, type=int,
+    parser.add_argument('--n_epochs_init', default=MuseCoEmbeddingGenerator.N_EPOCHS_INIT, type=int,
                         help='# of init training epochs')
-    parser.add_argument('--n_epochs', default=500, type=int,
+    parser.add_argument('--n_epochs', default=EmbeddingGenerator.N_EPOCHS, type=int,
                         help='# of training epochs')
-    parser.add_argument('--jackknife_percent', default=0.0, type=float,
+    parser.add_argument('--jackknife_percent', default=EmbeddingGenerator.JACKKNIFE_PERCENT, type=float,
                         help='Percentage of data to withhold from training'
                              'a value of 0.1 means to withhold 10 percent of the data')
     parser.add_argument('--mean_losses', action='store_true',
                         help='If set, use mean of losses otherwise sum')
-    parser.add_argument('--dropout', default=0.0, type=float,
+    parser.add_argument('--dropout', default=EmbeddingGenerator.DROPOUT, type=float,
                         help='Percentage to use fo dropout layers in neural network')
     parser.add_argument('--l2_norm', action='store_true',
                         help='If set, L2 normalize coembeddings')
     parser.add_argument('--fake_embedding', action='store_true',
                         help='If set, generate fake coembeddings')
+    parser.add_argument('--provenance',
+                        help='Path to file containing provenance '
+                             'information about input files in JSON format. '
+                             'This is required if none of embeddings directory contains '
+                             'ro-crate-metadata.json file.')
     parser.add_argument('--name',
                         help='Name of this run, needed for FAIRSCAPE. If '
                              'unset, name value from specified '
-                             'by --embeddings directories will be used')
+                             'by --embeddings directories or provenance file will be used')
     parser.add_argument('--organization_name',
                         help='Name of organization running this tool, needed '
                              'for FAIRSCAPE. If unset, organization name specified '
-                             'in --embedding directories will be used')
+                             'in --embedding directories or provenance file will be used')
     parser.add_argument('--project_name',
                         help='Name of project running this tool, needed for '
                              'FAIRSCAPE. If unset, project name specified '
-                             'in --embedding directories will be used')
+                             'in --embedding directories or provenance file will be used')
     parser.add_argument('--logconf', default=None,
                         help='Path to python logging configuration file in '
                              'this format: https://docs.python.org/3/library/'
@@ -144,6 +150,12 @@ def main(args):
         raise CellmapsCoEmbeddingError('Either --ppi_embeddingdir and --image_embeddingdir, '
                                        'or --embeddings are required')
 
+    if theargs.provenance is not None:
+        with open(theargs.provenance, 'r') as f:
+            json_prov = json.load(f)
+    else:
+        json_prov = None
+
     try:
         logutils.setup_cmd_logging(theargs)
         gen = EmbeddingGenerator()
@@ -175,7 +187,7 @@ def main(args):
                                                jackknife_percent=theargs.jackknife_percent,
                                                outdir=os.path.abspath(theargs.outdir),
                                                embeddings=theargs.embeddings,
-                                               embedding_names=theargs.embedding_names, 
+                                               embedding_names=theargs.embedding_names,
                                                mean_losses=theargs.mean_losses)
 
         inputdirs = gen.get_embedding_inputdirs()
@@ -186,7 +198,8 @@ def main(args):
                                   organization_name=theargs.organization_name,
                                   project_name=theargs.project_name,
                                   skip_logging=theargs.skip_logging,
-                                  input_data_dict=theargs.__dict__).run()
+                                  input_data_dict=theargs.__dict__,
+                                  provenance=json_prov).run()
     except Exception as e:
         logger.exception('Caught exception: ' + str(e))
         return 2